Overlap strings end alignment with mismatch/indels
Dynamic programming example of string matching/line up with errors and gaps.   Overlap specifically aligns end of one string with the end of another deciding where to start that overlap and scores it, selecting the best one.  In case of ties, options are shown at runtime by choosing from a list of ties by position.  Two datasets are stored, one with answer and score.